+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/PendingCertifiedProductDAO.java
@@ -3,7 +3,7 @@ package gov.healthit.chpl.dao;
 import java.util.List;
 
 import gov.healthit.chpl.dto.PendingCertifiedProductDTO;
-import gov.healthit.chpl.entity.PendingCertifiedProductEntity;
+import gov.healthit.chpl.entity.listing.pending.PendingCertifiedProductEntity;
 
 public interface PendingCertifiedProductDAO {