+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/CertifiedProduct.java
@@ -61,9 +61,9 @@ public class CertifiedProduct implements Serializable {
                     + dto.getProductCode() + "." + dto.getVersionCode() + "." + dto.getIcsCode() + "."
                     + dto.getAdditionalSoftwareCode() + "." + dto.getCertifiedDateCode());
         }
-        this.setLastModifiedDate(dto.getLastModifiedDate().getTime() + "");
+        this.setLastModifiedDate(dto.getLastModifiedDate() != null ? dto.getLastModifiedDate().getTime() + "" : "");
         this.edition = dto.getYear();
-        this.certificationDate = dto.getCertificationDate().getTime();
+        this.certificationDate = (dto.getCertificationDate() != null ? dto.getCertificationDate().getTime() : -1);
     }
 
     public Long getId() {