+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/entity/listing/pending/PendingCqmCriterionEntity.java
@@ -1,4 +1,4 @@
-package gov.healthit.chpl.entity;
+package gov.healthit.chpl.entity.listing.pending;
 
 import java.util.Date;
 import java.util.HashSet;
@@ -15,6 +15,8 @@ import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
+
+import gov.healthit.chpl.entity.CQMCriterionEntity;
 
 @Entity
 @Table(name = "pending_cqm_criterion")