+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/CertifiedProductController.java
@@ -52,7 +52,7 @@ import gov.healthit.chpl.dto.CertifiedProductDTO;
 import gov.healthit.chpl.dto.CertifiedProductDetailsDTO;
 import gov.healthit.chpl.dto.PendingCertifiedProductDTO;
 import gov.healthit.chpl.entity.CertificationStatusType;
-import gov.healthit.chpl.entity.PendingCertifiedProductEntity;
+import gov.healthit.chpl.entity.listing.pending.PendingCertifiedProductEntity;
 import gov.healthit.chpl.manager.ActivityManager;
 import gov.healthit.chpl.manager.CertificationBodyManager;
 import gov.healthit.chpl.manager.CertifiedProductDetailsManager;
@@ -396,8 +396,18 @@ public class CertifiedProductController {
             notes = "Accepts a CSV file with very specific fields to create pending certified products. "
                     + " The user uploading the file must have ROLE_ACB_ADMIN or ROLE_ACB_STAFF "
                     + " and administrative authority on the ACB(s) specified in the file.")
+    @RequestMapping(value = "/upload", method = RequestMethod.PUT, produces = "application/json; charset=utf-8")
+    public ResponseEntity<PendingCertifiedProductResults> uploadAsPut(@RequestParam("file") MultipartFile file)
+            throws ValidationException, MaxUploadSizeExceededException {
+        return upload(file);
+    }
+
+    @ApiOperation(value = "Upload a file with certified products",
+            notes = "Accepts a CSV file with very specific fields to create pending certified products. "
+                    + " The user uploading the file must have ROLE_ACB_ADMIN or ROLE_ACB_STAFF "
+                    + " and administrative authority on the ACB(s) specified in the file.")
     @RequestMapping(value = "/upload", method = RequestMethod.POST, produces = "application/json; charset=utf-8")
-    public @ResponseBody PendingCertifiedProductResults upload(@RequestParam("file") MultipartFile file)
+    public ResponseEntity<PendingCertifiedProductResults> upload(@RequestParam("file") MultipartFile file)
             throws ValidationException, MaxUploadSizeExceededException {
         if (file.isEmpty()) {
             throw new ValidationException("You cannot upload an empty file!");
@@ -407,7 +417,7 @@ public class CertifiedProductController {
                 && !file.getContentType().equalsIgnoreCase("application/vnd.ms-excel")) {
             throw new ValidationException("File must be a CSV document.");
         }
-
+        HttpHeaders responseHeaders = new HttpHeaders();
         List<PendingCertifiedProductDetails> uploadedProducts = new ArrayList<PendingCertifiedProductDetails>();
 
         BufferedReader reader = null;
@@ -456,6 +466,10 @@ public class CertifiedProductController {
                                     try {
                                         CertifiedProductUploadHandler handler = uploadHandlerFactory.getHandler(heading,
                                                 rows);
+                                        if(handler.getUploadTemplateVersion() != null && 
+                                                handler.getUploadTemplateVersion().getDeprecated() == Boolean.TRUE) {
+                                            responseHeaders.set(HttpHeaders.WARNING, "299 - \"Deprecated upload template\"");
+                                        }
                                         PendingCertifiedProductEntity pendingCp = handler.handle();
                                         cpsToAdd.add(pendingCp);
                                     } catch (final InvalidArgumentsException ex) {
@@ -476,6 +490,10 @@ public class CertifiedProductController {
                 if (i == records.size() - 1 && !rows.isEmpty()) {
                     try {
                         CertifiedProductUploadHandler handler = uploadHandlerFactory.getHandler(heading, rows);
+                        if(handler.getUploadTemplateVersion() != null && 
+                                handler.getUploadTemplateVersion().getDeprecated() == Boolean.TRUE) {
+                            responseHeaders.set(HttpHeaders.WARNING, "299 - \"Deprecated upload template\"");
+                        }
                         PendingCertifiedProductEntity pendingCp = handler.handle();
                         cpsToAdd.add(pendingCp);
                     } catch (final InvalidArgumentsException ex) {
@@ -527,8 +545,9 @@ public class CertifiedProductController {
             }
         }
 
+        
         PendingCertifiedProductResults results = new PendingCertifiedProductResults();
         results.getPendingCertifiedProducts().addAll(uploadedProducts);
-        return results;
+        return new ResponseEntity<PendingCertifiedProductResults>(results, responseHeaders, HttpStatus.OK);
     }
 }